+++ locksmith/__tests__/signatureValidation/signPersonal.test.ts
@@ -34,7 +34,7 @@ const body = {
       { name: 'address', type: 'address' },
     ],
   },
-  domain: { name: 'Unlock Dashboard', version: '1', chainId: 1984 },
+  domain: { name: 'Unlock Dashboard', version: '1', chainId: 1337 },
   primaryType: 'Lock',
   message: {
     lock: {
@@ -206,7 +206,7 @@ describe('Signature Validation Middleware', () => {
               { name: 'address', type: 'address' },
             ],
           },
-          domain: { name: 'Unlock Dashboard', version: '1', chainId: 1984 },
+          domain: { name: 'Unlock Dashboard', version: '1', chainId: 1337 },
           primaryType: 'Lock',
           message: {
             lock: {