+++ locksmith/src/utils/decoyUser.ts
@@ -1,4 +1,4 @@
-import { encrypt } from 'ethers/utils/secret-storage'
+import { ethers } from 'ethers'
 import randomWords from 'random-words'
 
 export class DecoyUser {
@@ -7,10 +7,7 @@ export class DecoyUser {
   }
 
   async encryptedPrivateKey() {
-    return encrypt(
-      '0x4c0883a69102937d6231471b5dbb6204fe5129617082792ae468d01a3f362318',
-      randomWords(5).join(' ')
-    )
+    return ethers.Wallet.createRandom().encrypt(randomWords(5).join(' '))
   }
 }