+++ locksmith/src/utils/envLoader.js
@@ -10,6 +10,7 @@ const configPath = path.resolve(
   '..',
   `.env.${unlockEnv}.local`
 )
+// eslint-disable-next-line no-console
 console.log(`Loading config from ${configPath}`)
 
 dotenv.config({