+++ unlock-js/src/__tests__/setup/prepare-ganache-for-unlock.js
@@ -25,7 +25,7 @@ serverIsUp(host, port, 1000 /* every second */, 120 /* up to 2 minutes */)
   .then(async () => {
     // Instantiate the walletService
     const walletService = new WalletService({
-      1984: {
+      1337: {
         provider: providerURL,
       },
     })